Validaciones

+++ FransfordSystem/Models/Examen.cs
@@ -29,8 +29,8 @@ namespace FransfordSystem.Models
         [Display(Name = "Precio del examen")]
         [Range(0, 999.99)]
         [Required(ErrorMessage = "Este campo es obligatorio, ingrese cantidades entre 0 a 999.99")]
-        //[RegularExpression(@"^[0-9]{3}.[0-9]{2}$", ErrorMessage = "Ingrese una cantidad valida")]
-        public float? PrecioExamen { get; set; }
+        //[RegularExpression(@"^([0-9]{3}.[0-9]{2})$", ErrorMessage = "Ingrese una cantidad valida")]
+        public double? PrecioExamen { get; set; }
 
         [Display(Name = "Tipo de muestra")]
         [Required(ErrorMessage = "Este campo es obligatorio")]
